+++ examples/state.c
@@ -5,9 +5,9 @@
 #include <stdio.h>
 
 // clang-format off
-#define State_INTERFACE(FN, CTX)         \
-    FN(CTX, int,  get, void *self)       \
-    FN(CTX, void, set, void *self, int x)
+#define State_INTERFACE(OP, CTX)         \
+    OP(CTX,  int, get, void *self)       \
+    OP(CTX, void, set, void *self, int x)
 // clang-format on
 
 interface(State);
@@ -16,13 +16,9 @@ typedef struct {
     int x;
 } Num;
 
-int Num_State_get(void *self) {
-    return ((Num *)self)->x;
-}
+int Num_get(void *self) { return ((Num *)self)->x; }
 
-void Num_State_set(void *self, int x) {
-    ((Num *)self)->x = x;
-}
+void Num_set(void *self, int x) { ((Num *)self)->x = x; }
 
 impl(State, Num);