+++ tests/common.c
@@ -1,20 +1,16 @@
 #include "common.h"
 
-const char *a1_impl(int x, double *restrict y) {
+const char *foo1_impl(int x, double *restrict y) {
     (void)x;
     (void)y;
     return (const char *)NULL;
 }
 
-float b1_impl(long long x) {
+float bar1_impl(long long x) {
     (void)x;
     return 12.141f;
 }
 
-const char *a2_impl(int x, double *restrict y) {
-    return a1_impl(x, y);
-}
+const char *foo2_impl(int x, double *restrict y) { return foo1_impl(x, y); }
 
-float b2_impl(long long x) {
-    return b1_impl(x);
-}
+float bar2_impl(long long x) { return bar1_impl(x); }