+++ tests/common.h
@@ -9,23 +9,24 @@
 
 #define Marker_INTERFACE
 
-#define Foo1_INTERFACE(FN, CTX) FN(CTX, const char *, a, int x, double *restrict y)
+#define Foo_INTERFACE(OP, CTX) OP(CTX, const char *, foo, int x, double *restrict y)
 
-#define Foo2_INTERFACE(FN, CTX)                                                                    \
-    FN(CTX, const char *, a, int x, double *restrict y)                                            \
-    FN(CTX, float, b, long long x)
+#define Bar_INTERFACE(OP, CTX)                                                                     \
+    OP(CTX, const char *, foo, int x, double *restrict y)                                          \
+    OP(CTX, float, bar, long long x)
 
 interface(Marker);
-interface(Foo1);
-interface(Foo2);
+interface(Foo);
+interface(Bar);
 // } (Interfaces)
 
-const char *a1_impl(int x, double *restrict y);
-float b1_impl(long long x);
-const char *a2_impl(int x, double *restrict y);
-float b2_impl(long long x);
+const char *foo1_impl(int x, double *restrict y);
+float bar1_impl(long long x);
 
-typedef const char *(*AFnType)(int x, double *restrict y);
-typedef float (*BFnType)(long long x);
+const char *foo2_impl(int x, double *restrict y);
+float bar2_impl(long long x);
+
+typedef const char *(*FooOpType)(int x, double *restrict y);
+typedef float (*BarOpType)(long long x);
 
 #endif // INTERFACE99_TESTS_COMMON_H