+++ tests/extern_impl/impl.c
@@ -3,9 +3,9 @@
 #include "../common.h"
 #include "types.h"
 
-#define Foo1Impl_Foo1_a a1_impl
-externImpl(Foo1, Foo1Impl);
+#define FooImpl_foo foo1_impl
+externImpl(Foo, FooImpl);
 
-#define Foo2Impl_a a1_impl
-#define Foo2Impl_b b1_impl
-externImplPrimary(Foo2, Foo2Impl);
+#define BarImpl_foo foo1_impl
+#define BarImpl_bar bar1_impl
+externImpl(Bar, BarImpl);