+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+ - Remove `implPrimary`, `externImplPrimary` [**BC**].
+
 ## [0.4.0] - 2021-09-14
 
 ### Changed